refactor: content와 관련된 도메인 분리
