+++ back/src/test/java/com/baba/back/content/domain/content/ContentTest.java
@@ -1,7 +1,6 @@
-package com.baba.back.content.domain;
+package com.baba.back.content.domain.content;
 
 import com.baba.back.baby.domain.Baby;
-import com.baba.back.content.domain.content.Content;
 import java.time.LocalDate;
 import java.util.UUID;
 import org.assertj.core.api.Assertions;
